+++ core/src/main/java/org/nervos/appchain/tx/ManagedTransaction.java
@@ -1,6 +1,7 @@
 package org.nervos.appchain.tx;
 
 import java.io.IOException;
+import java.math.BigInteger;
 
 import org.nervos.appchain.protocol.AppChainj;
 import org.nervos.appchain.protocol.core.methods.response.TransactionReceipt;
@@ -25,7 +26,7 @@ public abstract class ManagedTransaction {
 
     protected TransactionReceipt send(
             String to, String data, long quota, String nonce,
-            long validUntilBlock, int version , int chainId, String value)
+            long validUntilBlock, int version , BigInteger chainId, String value)
             throws IOException, TransactionException {
         return transactionManager.executeTransaction(
                 to, data, quota, nonce, validUntilBlock, version, chainId, value);
